# app/controllers/VehiclesController.py
from fastapi import APIRouter, Depends, HTTPException, status
from sqlalchemy.orm import Session
from pydantic import BaseModel
from app.models.Money import Money
from app.models.InsertVehicle import InsertVehicle
from app.models.UpdateVehiclePatch import UpdateVehiclePatch

from ..HeaderSecurity import getAuthData, adminAuthData    
from ..services.VehiclesService import VehiclesService
from ..repositories.VehiclesRepository import VehiclesRepository
from ..DBConn import getDB
from .. import Logger
from ..models.Vehicle import Vehicle
from ..models.VehicleFilters import VehicleFilters
from app.Dependencies import getCacheClient

logger = Logger.createLogger(__name__)
router = APIRouter()


@router.get("/veiculos")
def getVehicles(
    filters: VehicleFilters = Depends(),
    offset: int = 0,
    limit: int = 20,
    authData: dict = Depends(getAuthData), 
    db: Session = Depends(getDB),
    cache = Depends(getCacheClient)
):

    dollarValueKey = cache.get("tinnova:dolar2real") 
    dollar2real = None if dollarValueKey == b'error' else  Money(float(dollarValueKey.decode('utf-8')))

    logger.info(f"Attempting to retrieve vehicles for user: {authData['user']}. Dolar {dollar2real} BRL")

    repository = VehiclesRepository(db)
    service = VehiclesService(repository)

    if  any([filters.minPreco, filters.maxPreco]):
        logger.info(f"getting vehicles with price range: minPreco={filters.minPreco}, maxPreco={filters.maxPreco}")

        minPrice = int(filters.minPreco*100)
        maxPrice = int(filters.maxPreco*100)
        vehicles: list[Vehicle] | None = service.getVehicleByPrice(minPrice, maxPrice, offset, limit)
    
    if  any([filters.ano, filters.marca, filters.cor ]): # and not any([filters.minPreco, filters.maxPreco]):
        logger.info(f"getting vehicles with filters: ano={filters.ano}, marca={filters.marca}, cor={filters.cor}")
        vehicles: list[Vehicle] | None = service.getVehiclesByBrandYearColor(
            year=filters.ano,
            brand=filters.marca,
            color=filters.cor,
            offset=offset,
            limit=limit
        )


    if not any([filters.ano, filters.marca, filters.cor, filters.minPreco, filters.maxPreco]):
        logger.info("No filters provided, fetching all vehicles")


        vehicles: list[Vehicle] | None = service.getAllVehicles()

    for v in vehicles:
        real = Money((float(v.price_dolar_cents) / 100) * float(dollar2real)) if dollar2real else None
        dollar = Money(float(v.price_dolar_cents) / 100) if v.price_dolar_cents else None
        v.price_in_real = real if real else None
        v.price_in_dollar = dollar if dollar else None

    return {"vehicles": vehicles}

@router.get("/veiculos/{uuid}")
def getVehicleByUUID(uuid: str, data: dict = Depends(getAuthData), db: Session = Depends(getDB)):
    logger.info(f"Attempting to retrieve vehicle with UUID: {uuid} for user: {data['user']}")

    repository = VehiclesRepository(db)
    service = VehiclesService(repository)
    vehicle: Vehicle | None = service.getVehicleByUUID(uuid)

    if vehicle is None:
        logger.warning(f"Vehicle with UUID: {uuid} not found")
        raise HTTPException(
            status_code=status.HTTP_404_NOT_FOUND,
            detail="Vehicle not found!"
        )
    
    return {"vehicle": vehicle}

@router.get("/veiculos/relatorios/por-marca")
def getVehicleReportByBrand(data: dict = Depends(getAuthData), db: Session = Depends(getDB)):
    logger.info(f"Attempting to retrieve vehicle report by brand for user: {data['user']}")
    repository = VehiclesRepository(db)
    service = VehiclesService(repository)
    report = service.getVehicleReportByBrand()
    return {"report": report}

@router.post("/veiculos")
def createVehicle(vehicleData: InsertVehicle, data: dict = Depends(adminAuthData), db: Session = Depends(getDB)):
    logger.info(f"Attempting to create vehicle with data: {vehicleData} for admin user: {data['user']}")
    repository = VehiclesRepository(db)
    service = VehiclesService(repository)
    success = service.createVehicle(vehicleData)

    if not success:
        logger.warning(f"Failed to create vehicle with data: {vehicleData}")
        raise HTTPException(
            status_code=status.HTTP_400_BAD_REQUEST,
            detail="Failed to create vehicle! {}"
        )
    logger.info(f"Vehicle created successfully with data: {vehicleData}")
    return {"message": "Vehicle created successfully!"}


@router.put("/veiculos/{uuid}")
def updateCompleteVehicle(uuid: str, vehicleData: InsertVehicle, data: dict = Depends(adminAuthData), db: Session = Depends(getDB)):
    logger.info(f"Attempting to update vehicle with UUID: {uuid} using data: {vehicleData} for admin user: {data['user']}")
    repository = VehiclesRepository(db)
    service = VehiclesService(repository)
    success = service.updateCompleteVehicle(uuid, vehicleData)

    if not success:
        logger.warning(f"Failed to update vehicle with UUID: {uuid} using data: {vehicleData}")
        raise HTTPException(
            status_code=status.HTTP_400_BAD_REQUEST,
            detail="Failed to update vehicle! {}"
        )
    logger.info(f"Vehicle with UUID: {uuid} updated successfully using data: {vehicleData}")
    return {"message": "Vehicle updated successfully!"}

@router.delete("/veiculos/{uuid}")
def deleteVehicle(uuid: str, hardDelete: bool = False, data: dict = Depends(adminAuthData), db: Session = Depends(getDB)):
    logger.info(f"Attempting to delete vehicle with UUID: {uuid} for admin user: {data['user']}. Hard delete: {hardDelete}")
    repository = VehiclesRepository(db)
    service = VehiclesService(repository)
    success = service.deleteVehicle(uuid, hardDelete)

    if not success:
        logger.warning(f"Failed to delete vehicle with UUID: {uuid}. Hard delete: {hardDelete}")
        raise HTTPException(
            status_code=status.HTTP_400_BAD_REQUEST,
            detail="Failed to delete vehicle! {}"
        )
    logger.info(f"Vehicle with UUID: {uuid} deleted successfully. Hard delete: {hardDelete}")
    return {"message": "Vehicle deleted successfully!"}

@router.patch("/veiculos/{uuid}")
def updatePatchVehicle(uuid: str, updateData: UpdateVehiclePatch, data: dict = Depends(adminAuthData), db: Session = Depends(getDB)):
    logger.info(f"Attempting to partially update vehicle with UUID: {uuid} using data: {updateData} for admin user: {data['user']}")
    repository = VehiclesRepository(db)
    service = VehiclesService(repository)
    success = service.updatePatchVehicle(uuid, updateData)

    if not success:
        logger.warning(f"Failed to partially update vehicle with UUID: {uuid} using data: {updateData}")
        raise HTTPException(
            status_code=status.HTTP_400_BAD_REQUEST,
            detail="Failed to partially update vehicle! {}"
        )
    logger.info(f"Vehicle with UUID: {uuid} partially updated successfully using data: {updateData}")
    return {"message": "Vehicle partially updated successfully!"}