# documents/swagger.yaml
openapi: 3.1.0
info:
  title: Desafio Tinnova API
  version: 0.1.0
paths:
  /auth/login:
    post:
      tags:
        - Auth
      summary: Login
      operationId: login_auth_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /veiculos:
    get:
      summary: Getvehicles
      operationId: getVehicles_veiculos_get
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
            title: Offset
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
            title: Limit
        - name: ano
          in: query
          required: false
          schema:
            anyOf:
              - type: integer
              - type: 'null'
            title: Ano
        - name: marca
          in: query
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            title: Marca
        - name: cor
          in: query
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            title: Cor
        - name: minPreco
          in: query
          required: false
          schema:
            anyOf:
              - type: number
              - type: string
              - type: 'null'
            title: Minpreco
        - name: maxPreco
          in: query
          required: false
          schema:
            anyOf:
              - type: number
              - type: string
              - type: 'null'
            title: Maxpreco
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      summary: Createvehicle
      operationId: createVehicle_veiculos_post
      security:
        - OAuth2PasswordBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InsertVehicle'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /veiculos/{uuid}:
    get:
      summary: Getvehiclebyuuid
      operationId: getVehicleByUUID_veiculos__uuid__get
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: uuid
          in: path
          required: true
          schema:
            type: string
            title: Uuid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      summary: Updatecompletevehicle
      operationId: updateCompleteVehicle_veiculos__uuid__put
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: uuid
          in: path
          required: true
          schema:
            type: string
            title: Uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InsertVehicle'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      summary: Deletevehicle
      operationId: deleteVehicle_veiculos__uuid__delete
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: uuid
          in: path
          required: true
          schema:
            type: string
            title: Uuid
        - name: hardDelete
          in: query
          required: false
          schema:
            type: boolean
            default: false
            title: Harddelete
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      summary: Updatepatchvehicle
      operationId: updatePatchVehicle_veiculos__uuid__patch
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: uuid
          in: path
          required: true
          schema:
            type: string
            title: Uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateVehiclePatch'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /veiculos/relatorios/por-marca:
    get:
      summary: Getvehiclereportbybrand
      operationId: getVehicleReportByBrand_veiculos_relatorios_por_marca_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
        - OAuth2PasswordBearer: []
  /:
    get:
      summary: Root
      operationId: root__get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
components:
  schemas:
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    InsertVehicle:
      properties:
        complement:
          type: string
          title: Complement
        year:
          type: integer
          title: Year
        color:
          type: string
          title: Color
        plate:
          type: string
          title: Plate
        brand_id:
          type: integer
          title: Brand Id
        price:
          type: integer
          title: Price
      type: object
      required:
        - complement
        - year
        - color
        - plate
        - brand_id
        - price
      title: InsertVehicle
    LoginRequest:
      properties:
        username:
          type: string
          title: Username
        password:
          type: string
          title: Password
      type: object
      required:
        - username
        - password
      title: LoginRequest
    UpdateVehiclePatch:
      properties:
        brand_id:
          anyOf:
            - type: integer
            - type: 'null'
          title: Brand Id
        complement:
          anyOf:
            - type: string
            - type: 'null'
          title: Complement
        year:
          anyOf:
            - type: integer
            - type: 'null'
          title: Year
        color:
          anyOf:
            - type: string
            - type: 'null'
          title: Color
        price:
          anyOf:
            - type: number
            - type: 'null'
          title: Price
        plate:
          anyOf:
            - type: string
            - type: 'null'
          title: Plate
      type: object
      title: UpdateVehiclePatch
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
              - type: string
              - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
        - loc
        - msg
        - type
      title: ValidationError
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: login